requirements.txt
apache-airflow==2.7.3
pandas
requests
psycopg2-binary
